Inital upload of files.
This is the inital beta upload of the most recent decent working copy.
All of the files needed are included including pre-built copies of the
ANE. The bbmsp-ane folder is the NDK library, the BBM-ane folder is the
AIR library, and the BBM_ANE_Test folder is the test application I used.
In this version the Registration and Download Invites are working. The
images can be loaded and passed to the ANE, but I have not been able to
retrieve them and since I have not finished the Profile or Contact
portions, so I don't know if the passed data is still valid.

When using the test application, you must click Status first before
clicking Reg UUID. I assume that if you keep my same UUID then we will
be able to see each other in BBM once I get the contact part down. My
BBMID for my DevAlpha is CMY2 and the PIN is 29D9370E.