Fixed Issue with image loading.
I fixed the issues with loading of images to the ANE. I was passing the
wrong number of arguments to the function, and not releasing the
ByteArray upon completion. Image loading now works. Still working on
getting the image back from the ANE if needed and notifying AS of
completion. Right now you can use a work around of counting the number
of images in the map as is done in the sample app, but this will require
that you don't upload an image until the previous image is complete.