Added functionality
+Finshed with the BBMSPUtil classes tro handle image loading. Image
loading is now working completely including callbacks to notify AS when
an image has been loaded to and retrieved from the ANE.
+Added in the BBMSPContext class to allow users to check status of
access to facilities provided by BBMSP (platform access, download
invites, profile box access).
+Completed functions for the BBMSPUserProfile class. Profile is now
automatically retrieved once access has been granted. Can also push
updates to the profiles. Just need to add in the callbacks to alert AS
of the satutses.
+More work on BBMSPContactList but unable to completely test as don't
have another device to test on or any user to test with.