More work on user profile
Completed some more of the woork on user profiles. Everything is done
except the changing of the user profile picture. BBMSP only allows
pictures of size 32k so will need to incorporate one of the imaging
libraries to shrink down images passed to the ANE before setting as
profile picture.